requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
astropy==7.1.1
astropy-iers-data==0.2025.11.3.0.38.37
brotli==1.2.0
cachetools==6.2.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
dask==2025.11.0
fastapi==0.121.0
ffmpy==0.6.4
filelock==3.20.0
fsspec==2025.10.0
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
groovy==0.1.2
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
Jinja2==3.1.6
jplephem==2.23
locket==1.0.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.10.2
numpy==2.3.4
orjson==3.11.4
packaging==25.0
pandas==2.3.3
partd==1.4.2
patsy==1.0.2
pillow==11.3.0
plotly==6.4.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.10
pydantic_core==2.33.2
pydub==0.25.1
pyerfa==2.0.1.5
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.2.0
rsa==4.9.1
ruff==0.14.4
safehttpx==0.1.7
scipy==1.16.3
semantic-version==2.10.0
sgp4==2.25
shellingham==1.5.4
simpy==4.1.1
six==1.17.0
skyfield==1.53
sniffio==1.3.1
starlette==0.49.3
statsmodels==0.14.5
tomlkit==0.13.3
toolz==1.1.0
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
